requirements.txt
asgiref==3.7.1
async-timeout==4.0.2
Django==4.2.1
django-admin-volt==1.0.10
django-cors-headers==4.0.0
django-material-admin==1.8.6
django-redis==5.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
pytz==2023.3
redis==4.5.5
sqlparse==0.4.4
tzdata==2023.3
